safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFilterPrimitiveStandardAttributes-svgdom-height-prop-expected.txt

CONSOLE MESSAGE: line 47: TypeError: 'undefined' is not an object (evaluating 'gradientElement.height.baseVal')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'height' property of the SVGFilterPrimitiveStandardAttributes object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL gradientElement.height.baseVal.value should be 100. Threw exception TypeError: 'undefined' is not an object (evaluating 'gradientElement.height.baseVal')
